# frontend/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/cases": {
            "get": {
                "tags": [
                    "cases"
                ],
                "summary": "List Cases",
                "operationId": "list_cases_cases_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "enum": [
                                        "draft",
                                        "review",
                                        "published"
                                    ],
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Educators: filter by workflow status",
                            "title": "Status"
                        },
                        "description": "Educators: filter by workflow status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CaseResponse"
                                    },
                                    "title": "Response List Cases Cases Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "cases"
                ],
                "summary": "Create Case",
                "operationId": "create_case_cases_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cases/{id}": {
            "put": {
                "tags": [
                    "cases"
                ],
                "summary": "Update Case",
                "operationId": "update_case_cases__id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cases"
                ],
                "summary": "Delete Case",
                "operationId": "delete_case_cases__id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AcceptableAnswerResponse": {
                "properties": {
                    "field": {
                        "type": "string",
                        "title": "Field"
                    },
                    "answer": {
                        "type": "string",
                        "title": "Answer"
                    }
                },
                "type": "object",
                "required": [
                    "field",
                    "answer"
                ],
                "title": "AcceptableAnswerResponse"
            },
...
            "CaseResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "case_id": {
                        "type": "string",
                        "title": "Case Id"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "version": {
                        "type": "integer",
                        "title": "Version"
                    },
                    "created_by": {
                        "type": "string",
                        "title": "Created By"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "language": {
                        "type": "string",
                        "title": "Language"
                    },
                    "difficulty": {
                        "type": "string",
                        "title": "Difficulty"
                    },
                    "specialty": {
                        "type": "string",
                        "title": "Specialty"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags"
                    },
                    "age": {
                        "type": "integer",
                        "title": "Age"
                    },
                    "sex": {
                        "type": "string",
                        "title": "Sex"
                    },
                    "persona": {
                        "type": "string",
                        "title": "Persona"
                    },
                    "tone_presets": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tone Presets"
                    },
                    "chief_complaint": {
                        "type": "string",
                        "title": "Chief Complaint"
                    },
                    "history_of_present_illness": {
                        "type": "string",
                        "title": "History Of Present Illness"
                    },
                    "key_history_points": {
                        "$ref": "#/components/schemas/KeyHistoryPointsResponse"
                    },
                    "final_diagnosis": {
                        "type": "string",
                        "title": "Final Diagnosis"
                    },
                    "differential": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Differential"
                    },
                    "severity_or_stage": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Severity Or Stage"
                    },
                    "investigations": {
                        "$ref": "#/components/schemas/InvestigationsResponse"
                    },
                    "management": {
                        "$ref": "#/components/schemas/ManagementResponse"
                    },
                    "scoring": {
                        "$ref": "#/components/schemas/ScoringResponse"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "case_id",
                    "status",
                    "version",
                    "created_by",
                    "title",
                    "language",
                    "difficulty",
                    "specialty",
                    "tags",
                    "age",
                    "sex",
                    "persona",
                    "tone_presets",
                    "chief_complaint",
                    "history_of_present_illness",
                    "key_history_points",
                    "final_diagnosis",
                    "differential",
                    "severity_or_stage",
                    "investigations",
                    "management",
                    "scoring"
                ],
                "title": "CaseResponse"
            },
            "CreateCaseRequest": {
                "properties": {
                    "case_id": {
                        "type": "string",
                        "title": "Case Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "language": {
                        "type": "string",
                        "const": "en",
                        "title": "Language"
                    },
                    "difficulty": {
                        "type": "string",
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ],
                        "title": "Difficulty"
                    },
                    "specialty": {
                        "type": "string",
                        "title": "Specialty"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    },
                    "age": {
                        "type": "integer",
                        "title": "Age"
                    },
                    "sex": {
                        "type": "string",
                        "enum": [
                            "female",
                            "male",
                            "other"
                        ],
                        "title": "Sex"
                    },
                    "persona": {
                        "type": "string",
                        "title": "Persona"
                    },
                    "tone_presets": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tone Presets",
                        "default": []
                    },
                    "chief_complaint": {
                        "type": "string",
                        "title": "Chief Complaint"
                    },
                    "history_of_present_illness": {
                        "type": "string",
                        "title": "History Of Present Illness"
                    },
                    "key_history_points": {
                        "$ref": "#/components/schemas/KeyHistoryPointsRequest"
                    },
                    "final_diagnosis": {
                        "type": "string",
                        "title": "Final Diagnosis"
                    },
                    "differential": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Differential",
                        "default": []
                    },
                    "severity_or_stage": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Severity Or Stage"
                    },
                    "investigations": {
                        "$ref": "#/components/schemas/InvestigationsRequest"
                    },
                    "management": {
                        "$ref": "#/components/schemas/ManagementRequest"
                    },
                    "scoring": {
                        "$ref": "#/components/schemas/ScoringRequest"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "review",
                            "published"
                        ],
                        "title": "Status",
                        "default": "draft"
                    }
                },
                "type": "object",
                "required": [
                    "case_id",
                    "title",
                    "language",
                    "difficulty",
                    "specialty",
                    "age",
                    "sex",
                    "persona",
                    "chief_complaint",
                    "history_of_present_illness",
                    "key_history_points",
                    "final_diagnosis",
                    "investigations",
                    "management",
                    "scoring"
                ],
                "title": "CreateCaseRequest"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "InvestigationsResponse": {
                "properties": {
                    "catalog_hints": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Catalog Hints"
                    },
                    "expected": {
                        "$ref": "#/components/schemas/ExpectedTestsResponse"
                    },
                    "results": {
                        "items": {
                            "$ref": "#/components/schemas/InvestigationResultResponse"
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "catalog_hints",
                    "expected",
                    "results"
                ],
                "title": "InvestigationsResponse"
            },
...
            "KeyHistoryPointsResponse": {
                "properties": {
                    "must_ask": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Must Ask"
                    },
                    "nice_to_ask": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Nice To Ask"
                    },
                    "red_flags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Red Flags"
                    }
                },
                "type": "object",
                "required": [
                    "must_ask",
                    "nice_to_ask",
                    "red_flags"
                ],
                "title": "KeyHistoryPointsResponse"
            },
...
            "ManagementResponse": {
                "properties": {
                    "diagnostic_plan": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Diagnostic Plan"
                    },
                    "treatment_plan": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Treatment Plan"
                    },
                    "contraindications": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Contraindications"
                    },
                    "follow_up": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Follow Up"
                    }
                },
                "type": "object",
                "required": [
                    "diagnostic_plan",
                    "treatment_plan",
                    "contraindications",
                    "follow_up"
                ],
                "title": "ManagementResponse"
            },
...
            "ScoringResponse": {
                "properties": {
                    "weight_diagnosis": {
                        "type": "number",
                        "title": "Weight Diagnosis"
                    },
                    "weight_diagnostics": {
                        "type": "number",
                        "title": "Weight Diagnostics"
                    },
                    "weight_treatment": {
                        "type": "number",
                        "title": "Weight Treatment"
                    },
                    "weight_safety": {
                        "type": "number",
                        "title": "Weight Safety"
                    },
                    "acceptable_answers": {
                        "items": {
                            "$ref": "#/components/schemas/AcceptableAnswerResponse"
                        },
                        "type": "array",
                        "title": "Acceptable Answers"
                    },
                    "critical_safety_errors": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Critical Safety Errors"
                    }
                },
                "type": "object",
                "required": [
                    "weight_diagnosis",
                    "weight_diagnostics",
                    "weight_treatment",
                    "weight_safety",
                    "acceptable_answers",
                    "critical_safety_errors"
                ],
                "title": "ScoringResponse"
            },
...
            "UpdateCaseRequest": {
                "properties": {
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title"
                    },
                    "language": {
                        "anyOf": [
                            {
                                "type": "string",
                                "const": "en"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Language"
                    },
                    "difficulty": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "easy",
                                    "medium",
                                    "hard"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Difficulty"
                    },
                    "specialty": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Specialty"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags"
                    },
                    "age": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Age"
                    },
                    "sex": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "female",
                                    "male",
                                    "other"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Sex"
                    },
                    "persona": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Persona"
                    },
                    "tone_presets": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tone Presets"
                    },
                    "chief_complaint": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chief Complaint"
                    },
                    "history_of_present_illness": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "History Of Present Illness"
                    },
                    "key_history_points": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/KeyHistoryPointsRequest"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "final_diagnosis": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Final Diagnosis"
                    },
                    "differential": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Differential"
                    },
                    "severity_or_stage": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Severity Or Stage"
                    },
                    "investigations": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/InvestigationsRequest"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "management": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ManagementRequest"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "scoring": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ScoringRequest"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "status": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "draft",
                                    "review",
                                    "published"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Status"
                    }
                },
                "type": "object",
                "title": "UpdateCaseRequest"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/auth/login"
                    }
                }
            }
        }
    }
}